Added reselect library to cart-icon and cart-dropDown
Reselect slices a part of redux state so that we only use part of state that is required for that particular component while rendering it. This makes sure that even if the state has changed but the part of state corresponding to this component(which was sliced using reselect) has not changed, then it wont be re-rendered.